Crappy implementation of RAM saving
Persist extram to disk every 2 seconds because I was too tired to
implement proper caching/backoff and just wanted to get it working.

It works, but I definitely need to re-write bits of this, re-arrange
the code a bit, and delete the dead bits from when I used serialised
byte streams.

+++ src/test/java/com/kopieczek/gamble/savefiles/TestHashMapDb.java
@@ -10,8 +10,8 @@ import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
 
 public class TestHashMapDb {
-    private static byte[] ARRAY_1 = new byte[] {-10, 3, 127};
-    private static byte[] ARRAY_2 = new byte[] {100, 1, 2, 3, -127, 127, 19};
+    private static int[] ARRAY_1 = new int[] {-10, 3, 127};
+    private static int[] ARRAY_2 = new int[] {100, 1, 2, 3, -127, 127, 19};
 
     @Test
     public void test_open_empty_file_succeeds() throws Exception {
@@ -23,7 +23,7 @@ public class TestHashMapDb {
     public void test_put_and_retrieve_one_array() throws Exception {
         HashMapDb db = getTempDb();
         db.put("foo", ARRAY_1);
-        byte[] actual = db.get("foo").get();
+        int[] actual = db.get("foo").get();
         assertArrayEquals(ARRAY_1, actual);
     }
